+++ core/runner/src/NanoService.ts
@@ -1,8 +1,10 @@
 import { type ConfigContext, type Context, NodeBase, type ResponseContext } from "@nanoservice-ts/shared";
 import type ParamsDictionary from "@nanoservice-ts/shared/dist/types/ParamsDictionary";
+import { metrics } from "@opentelemetry/api";
 import { type Schema, type ValidationError, Validator } from "jsonschema";
 import _ from "lodash";
 import type { INanoServiceResponse } from "./NanoServiceResponse";
+import type RunnerNode from "./RunnerNode";
 import type JsonLikeObject from "./types/JsonLikeObject";
 
 export default abstract class NanoService extends NodeBase {
@@ -31,6 +33,16 @@ export default abstract class NanoService extends NodeBase {
 
 	public async run(ctx: Context): Promise<ResponseContext> {
 		const response: ResponseContext = { success: true, data: {}, error: null };
+		const defaultMeter = metrics.getMeter("default");
+		const start = performance.now();
+
+		const node_execution = defaultMeter.createCounter(`node:${this.name}`, {
+			description: "Node requests",
+		});
+
+		const node_time = defaultMeter.createGauge(`node:${this.name}:time`, {
+			description: "Node elapsed time",
+		});
 
 		const config = _.cloneDeep(ctx.config) as ConfigContext;
 		let opts: JsonLikeObject = (config as JsonLikeObject)[this.name] as unknown as JsonLikeObject;
@@ -48,6 +60,27 @@ export default abstract class NanoService extends NodeBase {
 			// Process node custom logic
 			const result = await this.handle(ctx, inputs as JsonLikeObject);
 			this.v.validate(result, this.outputSchema);
+			const end = performance.now();
+
+			node_execution.add(1, {
+				pid: process.pid,
+				env: process.env.NODE_ENV,
+				workflow_request_id: `${ctx.id}`,
+				workflow_runner_path: `${ctx.workflow_path}`,
+				workflow_runner_name: `${ctx.workflow_name}`,
+				node_name: `${this.name}`,
+				node: (this as unknown as RunnerNode).node,
+			});
+
+			node_time.record(end - start, {
+				pid: process.pid,
+				env: process.env.NODE_ENV,
+				workflow_request_id: `${ctx.id}`,
+				workflow_runner_path: `${ctx.workflow_path}`,
+				workflow_runner_name: `${ctx.workflow_name}`,
+				node_name: `${this.name}`,
+				node: (this as unknown as RunnerNode).node,
+			});
 
 			response.data = result;
 		} catch (error: unknown) {