+++ core/runner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nanoservice/runner
 
+## 0.1.9
+
+### Patch Changes
+
+- Improved and extended the open telemetry feature
+- Updated dependencies
+  - @nanoservice-ts/shared@0.0.3
+
 ## 0.1.8
 
 ### Patch Changes